# api/docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Сервис бронирования переговорок с автоматической генерацией слотов
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Meeting Room Booking Service API
  version: "1.0"
paths:
  /bookings/{bookingId}/cancel:
    post:
      consumes:
      - application/json
      description: Отменяет бронь. Идемпотентно. Только владелец брони.
      parameters:
      - description: ID брони
        in: path
        name: bookingId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: booking
          schema:
            additionalProperties: true
            type: object
        "401":
          description: error
          schema:
            additionalProperties: true
            type: object
        "403":
          description: error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Отменить бронь
      tags:
      - Bookings
  /bookings/create:
    post:
      consumes:
      - application/json
      description: Создаёт бронь на слот. Только user. userId берётся из JWT.
      parameters:
      - description: Данные брони
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: booking
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: error
          schema:
            additionalProperties: true
            type: object
        "403":
          description: error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: error
          schema:
            additionalProperties: true
            type: object
        "409":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать бронь
      tags:
      - Bookings
  /bookings/list:
    get:
      consumes:
      - application/json
      description: Возвращает список всех броней с пагинацией. Только admin.
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Размер страницы
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: bookings, pagination
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: error
          schema:
            additionalProperties: true
            type: object
        "403":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Список всех броней (admin)
      tags:
      - Bookings
  /bookings/my:
    get:
      consumes:
      - application/json
      description: Возвращает список активных броней текущего пользователя на будущие
        слоты.
      produces:
      - application/json
      responses:
        "200":
          description: bookings
          schema:
            additionalProperties: true
            type: object
        "401":
          description: error
          schema:
            additionalProperties: true
            type: object
        "403":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Мои брони
      tags:
      - Bookings
  /dummyLogin:
    post:
      consumes:
      - application/json
      description: Выдаёт тестовый JWT для указанной роли (admin/user). Для каждой
        роли возвращается фиксированный UUID.
      parameters:
      - description: Роль пользователя
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Получить тестовый JWT
      tags:
      - Auth
  /rooms/{roomId}/schedule/create:
    post:
      consumes:
      - application/json
      description: Создаёт расписание для переговорки. Только admin, только один раз.
        Длительность слота 30 минут.
      parameters:
      - description: ID переговорки
        in: path
        name: roomId
        required: true
        type: string
      - description: Расписание
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: schedule
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: error
          schema:
            additionalProperties: true
            type: object
        "403":
          description: error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: error
          schema:
            additionalProperties: true
            type: object
        "409":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать расписание переговорки
      tags:
      - Schedules
  /rooms/{roomId}/slots/list:
    get:
      consumes:
      - application/json
      description: Возвращает слоты, не занятые активной бронью, для указанной переговорки
        на указанную дату.
      parameters:
      - description: ID переговорки
        in: path
        name: roomId
        required: true
        type: string
      - description: Дата в формате YYYY-MM-DD
        example: "2026-04-10"
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: slots
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Список доступных слотов
      tags:
      - Slots
  /rooms/create:
    post:
      consumes:
      - application/json
      description: Создаёт новую переговорку. Только для admin.
      parameters:
      - description: Данные комнаты
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: room
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: error
          schema:
            additionalProperties: true
            type: object
        "403":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать переговорку
      tags:
      - Rooms
  /rooms/list:
    get:
      consumes:
      - application/json
      description: Возвращает список всех переговорок. Доступно admin и user.
      produces:
      - application/json
      responses:
        "200":
          description: rooms
          schema:
            additionalProperties: true
            type: object
        "401":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Список переговорок
      tags:
      - Rooms
securityDefinitions:
  BearerAuth:
    description: Введите токен в формате "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"